Move stream code in http_stream.h/c
